Fixed a bug in the showcase, sorting layers fix
The 'Default' sorting layer is now really default for the new meshes.
Changed the position bug in the roguelike demo that was probably caused
by the grid-sync (the player was one tile too low in the y-axis). Put
the mesh and player in the correct sorting order in this demo aswell.